Fix VideoFile class name
Changed it back to the original since all other variants are gone
